implement proper request handlers
CherryPy's most straightforward model is that a handler (coded by
developer) exists only once, and responds to each request, but with the
same instance ("self") for each request.  It's not possible to store
contextual information in one method for retrieval in another, because
the same context is shared by all requests concurrently.  (It may be
possible to modify the "cherrypy.request" object, though it appears to
be a global, but doing so has no true relation to the handler itself.)

This cset modifies the cpy workflow to instantiate a new foo handler
for each request to /foo, for cleaner and more OO context management.
It cleans up the event logging already, and will provide other benefits
as we progress.